+++ src/api/Rulings.ts
@@ -1,6 +1,5 @@
 import Cached from "../util/Cached";
 import MagicQuerier, { List } from "../util/MagicQuerier";
-import {CollectorNumber} from "./Cards";
 
 export interface Ruling {
 	source: string;
@@ -16,7 +15,7 @@ class Rulings extends MagicQuerier {
 	}
 
 	@Cached
-	public async bySet (setCode: string, collectorNumber: CollectorNumber) {
+	public async bySet (setCode: string, collectorNumber: string | number) {
 		return (await this.query<List<Ruling>>(["cards", setCode, `${collectorNumber}`, "rulings"])).data;
 	}